refactor(session-store): extract message graph owner
