+++ src/renderer/src/stores/session-store.ts
@@ -12,46 +12,32 @@ import {
   type AcpMessageImage,
   type AcpTurnTokenUsage
 } from '../../../shared/acp'
-import {
-  DEFAULT_PERMISSION_PROFILE,
-  type PermissionProfileId
-} from '../../../shared/permission-profiles'
+import type { PermissionProfileId } from '../../../shared/permission-profiles'
 import {
   sanitizeMessageImages,
-  sanitizeToolActivity,
-  sanitizeActivityGroup,
-  type MessagePart,
   type PersistedActivityGroup,
   type PersistedArtifact,
-  type PersistedChatMessage,
   type PersistedChatSession,
-  type PersistedUploadedAttachment,
-  type PersistedToolActivity
+  type PersistedUploadedAttachment
 } from '../../../shared/session-persistence'
 import type { UpdateSessionArchiveRequest } from '../../../shared/session-persistence'
 import { isReportableRunFailure } from '../../../shared/run-error-classification'
-import { PENDING_UPLOAD_SESSION_ID } from '../../../shared/uploads'
+import { createPersistedUpload } from './session-store-message-graph-helpers'
 import {
-  createLinearConversationGraph,
-  activateConversationBranch,
-  ensureConversationRuntimeSegment,
-  forkEditedConversationMessage,
-  projectConversationMessage,
-  resolveActiveConversationActivities,
-  resolveActiveConversationMessages,
-  synchronizeActiveConversationActivities,
-  synchronizeActiveConversationMessages
-} from '../../../shared/conversation-graph'
+  createMessageId,
+  createSessionMessageGraphOwner,
+  createSortIndex,
+  synchronizeSessionGraph
+} from './session-store-message-graph-owner'
+import type {
+  AppendMessageResult,
+  SessionMessageGraphActions
+} from './session-store-message-graph-helpers'
 import {
   createInitialSessionState,
   createSessionPersistenceOwner,
   hydrateSession,
-  hydrateToolActivity,
-  stripTransientMessageState,
-  type ActiveRun,
   type ChatMessage,
-  type ChatMessageRole,
-  type ChatMessageStatus,
   type ChatSession,
   type SessionStatus,
   type SessionPersistenceActions,
@@ -75,58 +61,7 @@ export {
   type ToolActivityStatus
 } from './session-store-persistence-owner'
 
-type AppendUserMessageInput = {
-  sessionId: string
-  content: string
-  attachments?: PersistedUploadedAttachment[]
-  parts?: MessagePart[]
-  cwd?: string
-  projectId?: string
-  permissionProfile?: PermissionProfileId
-  agentFrameworkId?: PersistedChatSession['agentFrameworkId']
-  agentBackendId?: PersistedChatSession['agentBackendId']
-  agentModel?: string
-  isPending?: boolean
-  // Immutable Specialist UUID; written once on session creation and never changed after.
-  specialistId?: string
-}
-
-type AppendPendingUserMessageInput = {
-  content: string
-  attachments?: PersistedUploadedAttachment[]
-  parts?: MessagePart[]
-  cwd?: string
-  projectId?: string
-  permissionProfile?: PermissionProfileId
-  agentFrameworkId?: PersistedChatSession['agentFrameworkId']
-  agentBackendId?: PersistedChatSession['agentBackendId']
-  agentModel?: string
-  // Immutable Specialist UUID forwarded from the new-conversation draft picker.
-  specialistId?: string
-}
-
-// Starts a fresh pending Session from the selected source's visible active path. Runtime settings
-// may be supplied by the normal-send path; omitted values retain the source Session's setting.
-export type BranchInNewSessionInput = {
-  sourceSessionId: string
-  content: string
-  attachments?: PersistedUploadedAttachment[]
-  parts?: MessagePart[]
-  permissionProfile?: PermissionProfileId
-  agentFrameworkId?: PersistedChatSession['agentFrameworkId']
-  agentBackendId?: PersistedChatSession['agentBackendId']
-  agentModel?: string
-  // `undefined` inherits the source binding; `null` deliberately chooses the Main Agent.
-  specialistId?: string | null
-}
-
-type BindPendingSessionInput = {
-  pendingSessionId: string
-  sessionId: string
-  cwd?: string
-  agentFrameworkId?: PersistedChatSession['agentFrameworkId']
-  agentBackendId?: PersistedChatSession['agentBackendId']
-}
+export type { BranchInNewSessionInput } from './session-store-message-graph-helpers'
 
 type AppendAgentMessageChunkInput = {
   sessionId: string
@@ -177,34 +112,11 @@ type ReplaceMessageUploadsInput = {
   uploads: PersistedUploadedAttachment[]
 }
 
-type AppendMessageResult = {
-  sessionId: string
-  messageId: string
-}
-
-type AppendRoutedUserMessageInput = {
-  sessionId: string
-  messageId: string
-  eventId: string
-  content: string
-  createdAt: number
-  responseToMessageId?: string
-}
-
 type SessionStore = SessionStoreData &
-  SessionPersistenceActions & {
+  SessionPersistenceActions &
+  SessionMessageGraphActions & {
     selectSession: (sessionId: string) => void
     clearSelection: () => void
-    appendUserMessage: (input: AppendUserMessageInput) => AppendMessageResult | undefined
-    appendRoutedUserMessage: (
-      input: AppendRoutedUserMessageInput
-    ) => AppendMessageResult | undefined
-    appendPendingUserMessage: (
-      input: AppendPendingUserMessageInput
-    ) => AppendMessageResult | undefined
-    branchInNewSession: (input: BranchInNewSessionInput) => AppendMessageResult | undefined
-    bindPendingSession: (input: BindPendingSessionInput) => AppendMessageResult | undefined
-    clearPendingContextReplay: (sessionId: string, messageId: string) => void
     appendAgentMessageChunk: (
       input: AppendAgentMessageChunkInput
     ) => AppendMessageResult | undefined
@@ -235,9 +147,6 @@ type SessionStore = SessionStoreData &
       agentBackendId?: PersistedChatSession['agentBackendId']
     ) => void
     markDisconnected: (sessionId: string, reason?: string) => void
-    removeMessage: (sessionId: string, messageId: string) => void
-    truncateSessionFromMessage: (sessionId: string, messageId: string) => void
-    activateMessageBranch: (sessionId: string, branchId: string) => void
     setBranchSwitchBlocked: (sessionId: string, blocked: boolean) => void
     clearBranchContextReset: (sessionId: string) => void
     markSpecialistSwitchResetRequired: (sessionId: string) => void
@@ -273,79 +182,9 @@ type SessionStore = SessionStoreData &
     removeSessionsForProject: (projectId: string) => void
   }
 
-// Keeps renderer message ids unique across store mutations in this process.
-let messageSequence = 0
-let pendingSessionSequence = 0
-let timelineSequence = 0
-let conversationBranchSequence = 0
-let runtimeSegmentSequence = 0
 const ARTIFACT_ERROR_PREFIX = 'Generated file finalization failed'
 const CONVERSATION_GRAPH_SYNC_ERROR =
   'Conversation history could not be finalized safely. Restart the app to restore the last saved conversation state, then report this issue.'
-
-// Creates renderer-local message ids while session ids come from the runtime.
-const createMessageId = (): string => {
-  messageSequence += 1
-  return `message-${Date.now()}-${messageSequence}`
-}
-
-// Creates a renderer-only session id while the ACP session request is still pending.
-const createPendingSessionId = (): string => {
-  pendingSessionSequence += 1
-  return `pending-session-${Date.now()}-${pendingSessionSequence}`
-}
-
-const createSortIndex = (): number => {
-  timelineSequence += 1
-  return timelineSequence
-}
-
-const createConversationBranchId = (): string => {
-  conversationBranchSequence += 1
-  return `message-branch-${Date.now()}-${conversationBranchSequence}`
-}
-
-const createRuntimeSegmentId = (): string => {
-  runtimeSegmentSequence += 1
-  return `runtime-segment-${Date.now()}-${runtimeSegmentSequence}`
-}
-
-const synchronizeSessionGraph = (
-  session: ChatSession,
-  messages: ChatMessage[],
-  now: number,
-  frameworkId = session.agentFrameworkId ?? 'claude-code',
-  backendId = session.agentBackendId,
-  model = session.agentModel
-): NonNullable<PersistedChatSession['conversationGraph']> => {
-  const projection = messages.map(stripTransientMessageState)
-  const initial =
-    session.conversationGraph ??
-    createLinearConversationGraph({
-      sessionId: session.id,
-      messages: session.messages.map(stripTransientMessageState),
-      frameworkId: session.agentFrameworkId,
-      backendId: session.agentBackendId,
-      model: session.agentModel,
-      createdAt: session.createdAt,
-      updatedAt: session.updatedAt
-    })
-  const withSegment = ensureConversationRuntimeSegment(initial, {
-    id: createRuntimeSegmentId(),
-    frameworkId,
-    backendId,
-    model,
-    startedAt: now
-  })
-  const withMessages = synchronizeActiveConversationMessages(withSegment, projection, now)
-  const persistedActivities = (session.activities ?? [])
-    .map(sanitizeToolActivity)
-    .filter((activity): activity is PersistedToolActivity => Boolean(activity))
-  const persistedGroups = (session.activityGroups ?? [])
-    .map(sanitizeActivityGroup)
-    .filter((group): group is PersistedActivityGroup => Boolean(group))
-  return synchronizeActiveConversationActivities(withMessages, persistedActivities, persistedGroups)
-}
 
 const CLEARED_AGENT_RUN_STATE = {
   activeRun: undefined,
@@ -403,135 +242,6 @@ const completeOpenActivityGroups = (
 
   return completed && completed.length > 0 ? completed : undefined
 }
-
-// Derives a compact default title from the first user message.
-const createTitleFromMessage = (content: string): string => {
-  const normalizedTitle = content.replace(/\s+/g, ' ').trim()
-
-  return normalizedTitle.length > 48 ? `${normalizedTitle.slice(0, 48)}...` : normalizedTitle
-}
-
-// A branch's title identifies the new request, so unlike the sidebar's first-message default it
-// keeps the complete normalized prompt. Sidebar truncation remains a presentation concern.
-const createBranchTitleFromMessage = (content: string): string =>
-  content.replace(/\s+/g, ' ').trim()
-
-// Provides a readable conversation title when the first prompt contains only attachments.
-const createTitleFromUploads = (uploads: PersistedUploadedAttachment[]): string => {
-  if (uploads.length === 1) return `Attached ${uploads[0].originalName || uploads[0].name}`
-
-  return `Attached ${uploads.length} files`
-}
-
-// Copies only upload reference fields into store state, never file bytes or renderer extras.
-const createPersistedUpload = (
-  attachment: PersistedUploadedAttachment
-): PersistedUploadedAttachment => ({
-  id: attachment.id,
-  sessionId: attachment.sessionId,
-  name: attachment.name,
-  originalName: attachment.originalName,
-  mimeType: attachment.mimeType,
-  size: attachment.size,
-  versionId: attachment.versionId,
-  versionNumber: attachment.versionNumber,
-  createdAt: attachment.createdAt,
-  sha256: attachment.sha256 ?? attachment.checksum,
-  // A staged upload has no Version yet. Keep its main-issued path capability in renderer memory
-  // until finalizeSession publishes the immutable Version; the persistence bridge explicitly waits
-  // for that transition, so newly written Session JSON remains path-free.
-  ...(!attachment.versionId && attachment.sessionId === PENDING_UPLOAD_SESSION_ID && attachment.path
-    ? { path: attachment.path }
-    : {})
-})
-
-// A path-only legacy Upload still needs its source-owned capability long enough for the branch send
-// path to publish it as an immutable Version. The child remains pending until that reconciliation
-// finishes, so this path is transient and never reaches a newly written Session JSON.
-const copySnapshotUpload = (
-  attachment: PersistedUploadedAttachment
-): PersistedUploadedAttachment => ({
-  ...createPersistedUpload(attachment),
-  ...(!attachment.versionId && attachment.path ? { path: attachment.path } : {})
-})
-
-// Normalizes message timestamps, ids, stream linkage, and status.
-const createMessage = (
-  role: ChatMessageRole,
-  content: string,
-  status: ChatMessageStatus,
-  streamId?: string,
-  eventIds: string[] = [],
-  uploads: PersistedUploadedAttachment[] = [],
-  parts?: MessagePart[]
-): ChatMessage => {
-  const now = Date.now()
-  // Normalize upload references before attaching them to durable message state.
-  const persistedUploads = uploads.map(createPersistedUpload)
-
-  return {
-    id: createMessageId(),
-    role,
-    content,
-    status,
-    streamId,
-    eventIds,
-    uploads: persistedUploads.length > 0 ? persistedUploads : undefined,
-    // Structured mention segments drive the styled bubble; omit them when absent.
-    parts: parts && parts.length > 0 ? parts : undefined,
-    sortIndex: createSortIndex(),
-    createdAt: now,
-    updatedAt: now
-  }
-}
-
-// Copies the visible transcript without carrying event/stream correlations into the fresh runtime.
-const copySnapshotMessage = (message: PersistedChatMessage, sortIndex: number): ChatMessage => ({
-  ...message,
-  streamId: undefined,
-  eventIds: [],
-  artifactIds: message.artifactIds ? [...message.artifactIds] : undefined,
-  uploads: message.uploads?.map(copySnapshotUpload),
-  images: message.images?.map((image) => ({ ...image })),
-  parts: message.parts?.map((part) => ({ ...part })),
-  turnUsage: message.turnUsage ? { ...message.turnUsage } : undefined,
-  sortIndex
-})
-
-// Tool rows are rendered as historical transcript only; provider event ids must not be reused by a
-// new Session's runtime. The bounded payload is otherwise preserved for the existing detail views.
-const createSnapshotActivityId = (sessionId: string, id: string): string =>
-  `history:${sessionId}:${id}`
-
-const copySnapshotActivity = (activity: PersistedToolActivity, sessionId: string): ToolActivity =>
-  hydrateToolActivity({
-    ...activity,
-    id: createSnapshotActivityId(sessionId, activity.id),
-    activityGroupId: activity.activityGroupId
-      ? createSnapshotActivityId(sessionId, activity.activityGroupId)
-      : undefined,
-    eventIds: [],
-    toolContent: activity.toolContent ? [...activity.toolContent] : undefined,
-    toolLocations: activity.toolLocations?.map((location) => ({ ...location }))
-  })
-
-const copySnapshotActivityGroup = (
-  group: PersistedActivityGroup,
-  sessionId: string
-): PersistedActivityGroup => ({
-  ...group,
-  id: createSnapshotActivityId(sessionId, group.id),
-  activityIds: group.activityIds.map((id) => createSnapshotActivityId(sessionId, id))
-})
-
-const canBranchInNewSession = (session: ChatSession): boolean =>
-  !session.activeRun &&
-  session.status !== 'running' &&
-  session.status !== 'waiting-permission' &&
-  !session.fixLoopActive &&
-  !session.compacting &&
-  !session.branchSwitchBlocked &&
-  !session.conversationGraphSyncBlocked
 
 // Converts main-process artifact metadata into the compact persisted renderer reference shape.
 const createPersistedArtifact = (artifact: ArtifactFile): PersistedArtifact => {
@@ -780,381 +490,7 @@ export const useSessionStore = create<SessionStore>((set, get) => ({
     set({ selectedSessionId: undefined })
   },
 
-  // Projects a routed, already-durable user Message into the visible transcript. Unlike a new
-  // prompt this continues the current interaction and must not create another active run.
-  appendRoutedUserMessage: ({
-    sessionId,
-    messageId,
-    eventId,
-    content,
-    createdAt,
-    responseToMessageId
-  }) => {
-    const trimmedContent = content.trim()
-    const session = get().sessions.find((candidate) => candidate.id === sessionId)
-    if (!session || !trimmedContent) return undefined
-    if (session.messages.some((message) => message.id === messageId)) {
-      return { sessionId, messageId }
-    }
-
-    const matchingFeedbackIndex = session.messages.findIndex(
-      (message) =>
-        message.role === 'user' &&
-        message.content.trim() === trimmedContent &&
-        (message.id.startsWith('local-user-message-') ||
-          messageId.startsWith('local-user-message-'))
-    )
-    const matchingFeedback = session.messages[matchingFeedbackIndex]
-    const isLocalMessage = messageId.startsWith('local-user-message-')
-    if (
-      matchingFeedback &&
-      (!matchingFeedback.id.startsWith('local-user-message-') || isLocalMessage)
-    ) {
-      return { sessionId, messageId: matchingFeedback.id }
-    }
-
-    const message: ChatMessage = {
-      id: messageId,
-      role: 'user',
-      content: trimmedContent,
-      status: 'complete',
-      eventIds: [eventId],
-      sortIndex: createSortIndex(),
-      createdAt,
-      updatedAt: createdAt,
-      ...(responseToMessageId ? { responseToMessageId } : {})
-    }
-    const messages = matchingFeedback
-      ? session.messages.map((existing, index) =>
-          index === matchingFeedbackIndex
-            ? { ...message, sortIndex: matchingFeedback.sortIndex }
-            : existing
-        )
-      : [...session.messages, message]
-    set({
-      sessions: get().sessions.map((candidate) =>
-        candidate.id === sessionId
-          ? {
-              ...candidate,
-              status: candidate.activeRun ? 'running' : candidate.status,
-              messages,
-              conversationGraph: synchronizeSessionGraph(candidate, messages, createdAt),
-              updatedAt: Math.max(candidate.updatedAt, createdAt)
-            }
-          : candidate
-      )
-    })
-    return { sessionId, messageId }
-  },
-
-  // Appends a user prompt, creating the session if this is its first message.
-  appendUserMessage: ({
-    sessionId,
-    content,
-    attachments = [],
-    parts,
-    cwd,
-    projectId,
-    permissionProfile,
-    agentFrameworkId,
-    agentBackendId,
-    agentModel,
-    isPending,
-    specialistId
-  }) => {
-    const trimmedContent = content.trim()
-    const normalizedAgentBackendId = agentBackendId?.trim() || undefined
-    const normalizedAgentModel = agentModel?.trim() || undefined
-    const uploads = attachments.map(createPersistedUpload)
-
-    if (!sessionId || (!trimmedContent && uploads.length === 0)) return undefined
-
-    const state = get()
-    const existingSession = state.sessions.find((session) => session.id === sessionId)
-    const now = Date.now()
-    const userMessage = createMessage(
-      'user',
-      trimmedContent,
-      'complete',
-      undefined,
-      [],
-      uploads,
-      parts
-    )
-    const activeRun: ActiveRun = {
-      promptMessageId: userMessage.id,
-      startedAt: now
-    }
-
-    // Existing sessions keep their message history and restart their active run.
-    if (existingSession) {
-      const replayPromptIndex = existingSession.pendingContextReplayMessageId
-        ? existingSession.messages.findIndex(
-            (message) => message.id === existingSession.pendingContextReplayMessageId
-          )
-        : -1
-      const nextMessages =
-        replayPromptIndex >= 0
-          ? existingSession.messages.map((message, index) =>
-              index === replayPromptIndex ? userMessage : message
-            )
-          : [...existingSession.messages, userMessage]
-      set({
-        selectedSessionId: sessionId,
-        sessions: state.sessions.map((session) =>
-          session.id === sessionId
-            ? {
-                ...session,
-                status: 'running',
-                activeRun,
-                ...(session.isPending
-                  ? {
-                      agentFrameworkId,
-                      agentBackendId: normalizedAgentBackendId
-                    }
-                  : {}),
-                agentModel: normalizedAgentModel,
-                agentStatus: undefined,
-                error: undefined,
-                // Clear the prior failure's report flag alongside its text so a later internal error
-                // cannot inherit a stale `false` and wrongly hide its Report button.
-                errorReportable: undefined,
-                compacting: undefined,
-                messages: nextMessages,
-                pendingContextReplayMessageId: replayPromptIndex >= 0 ? userMessage.id : undefined,
-                conversationGraph: synchronizeSessionGraph(
-                  replayPromptIndex >= 0
-                    ? { ...session, messages: nextMessages, conversationGraph: undefined }
-                    : session,
-                  nextMessages,
-                  now,
-                  agentFrameworkId ?? session.agentFrameworkId ?? 'claude-code',
-                  normalizedAgentBackendId ?? session.agentBackendId,
-                  normalizedAgentModel
-                ),
-                updatedAt: now
-              }
-            : session
-        )
-      })
-    } else {
-      // New sessions use the caller-provided id, which may be a pending renderer id.
-      const newSession: ChatSession = {
-        id: sessionId,
-        projectId: projectId ?? '',
-        isPending: isPending ? true : undefined,
-        title: createTitleFromMessage(trimmedContent || createTitleFromUploads(uploads)),
-        cwd: cwd ?? '',
-        status: 'running',
-        permissionProfile: permissionProfile ?? DEFAULT_PERMISSION_PROFILE,
-        agentFrameworkId,
-        agentBackendId: normalizedAgentBackendId,
-        agentModel: normalizedAgentModel,
-        // Specialist UUID written once at creation; never changed after bind.
-        ...(specialistId ? { specialistId } : {}),
-        messages: [userMessage],
-        activeRun,
-        createdAt: now,
-        updatedAt: now
-      }
-      newSession.conversationGraph = synchronizeSessionGraph(
-        newSession,
-        newSession.messages,
-        now,
-        agentFrameworkId ?? 'claude-code',
-        normalizedAgentBackendId,
-        normalizedAgentModel
-      )
-
-      set({
-        selectedSessionId: sessionId,
-        sessions: [newSession, ...state.sessions]
-      })
-    }
-
-    return { sessionId, messageId: userMessage.id }
-  },
-
-  // Creates a visible local conversation before ACP returns the durable session id.
-  appendPendingUserMessage: ({
-    content,
-    attachments = [],
-    parts,
-    cwd,
-    projectId,
-    permissionProfile,
-    agentFrameworkId,
-    agentBackendId,
-    agentModel,
-    specialistId
-  }) => {
-    return get().appendUserMessage({
-      sessionId: createPendingSessionId(),
-      content,
-      attachments,
-      parts,
-      cwd,
-      projectId,
-      permissionProfile,
-      agentFrameworkId,
-      agentBackendId,
-      agentModel,
-      specialistId,
-      isPending: true
-    })
-  },
-
-  // Creates/selects a fresh pending Session from only the source's current visible Branch. The
-  // runtime binds this local id before replaying the bounded history into a new ACP Session.
-  branchInNewSession: ({
-    sourceSessionId,
-    content,
-    attachments = [],
-    parts,
-    permissionProfile,
-    agentFrameworkId,
-    agentBackendId,
-    agentModel,
-    specialistId
-  }) => {
-    const trimmedContent = content.trim()
-    const uploads = attachments.map(createPersistedUpload)
-    if (!sourceSessionId || (!trimmedContent && uploads.length === 0)) return undefined
-
-    const state = get()
-    const source = state.sessions.find((session) => session.id === sourceSessionId)
-    if (!source || !canBranchInNewSession(source)) return undefined
-
-    // `messages`/activities are the store's active-Branch compatibility projection. Reading this
-    // single source snapshot keeps message and activity relationships aligned without traversing
-    // inactive Graph siblings.
-    const sourceMessages = source.messages.map((message, index) => ({
-      message: stripTransientMessageState(message),
-      // Match the timeline projection fallback so exact-timestamp tool rows keep their position.
-      sortIndex: message.sortIndex ?? index
-    }))
-    const sourceActivities = (source.activities ?? [])
-      .map(sanitizeToolActivity)
-      .filter((activity): activity is PersistedToolActivity => Boolean(activity))
-    const sourceActivityGroups = (source.activityGroups ?? [])
-      .map(sanitizeActivityGroup)
-      .filter((group): group is PersistedActivityGroup => Boolean(group))
-
-    const now = Date.now()
-    const sessionId = createPendingSessionId()
-    const userMessage = createMessage(
-      'user',
-      trimmedContent,
-      'complete',
-      undefined,
-      [],
-      uploads,
-      parts
-    )
-    const messages = [
-      ...sourceMessages.map(({ message, sortIndex }) => copySnapshotMessage(message, sortIndex)),
-      userMessage
-    ]
-    const normalizedAgentBackendId = agentBackendId?.trim() || source.agentBackendId
-    const normalizedAgentModel = agentModel?.trim() || source.agentModel
-    const normalizedFrameworkId = agentFrameworkId ?? source.agentFrameworkId
-    const nextSpecialistId =
-      specialistId === undefined ? source.specialistId : specialistId || undefined
-    const newSession: ChatSession = {
-      id: sessionId,
-      projectId: source.projectId,
-      isPending: true,
-      title: trimmedContent
-        ? createBranchTitleFromMessage(trimmedContent)
-        : createTitleFromUploads(uploads),
-      cwd: source.cwd,
-      status: 'running',
-      permissionProfile:
-        permissionProfile ?? source.permissionProfile ?? DEFAULT_PERMISSION_PROFILE,
-      agentFrameworkId: normalizedFrameworkId,
-      agentBackendId: normalizedAgentBackendId,
-      agentModel: normalizedAgentModel,
-      ...(source.autoReviewEnabled !== undefined
-        ? { autoReviewEnabled: source.autoReviewEnabled }
-        : {}),
-      ...(source.enabledComputeHosts
-        ? { enabledComputeHosts: [...source.enabledComputeHosts] }
-        : {}),
-      ...(nextSpecialistId ? { specialistId: nextSpecialistId } : {}),
-      messages,
-      activities: sourceActivities.map((activity) => copySnapshotActivity(activity, sessionId)),
-      activityGroups: sourceActivityGroups.map((group) =>
-        copySnapshotActivityGroup(group, sessionId)
-      ),
-      pendingContextReplayMessageId: userMessage.id,
-      activeRun: { promptMessageId: userMessage.id, startedAt: now },
-      createdAt: now,
-      updatedAt: now
-    }
-    newSession.conversationGraph = synchronizeSessionGraph(
-      newSession,
-      messages,
-      now,
-      normalizedFrameworkId ?? 'claude-code',
-      normalizedAgentBackendId,
-      normalizedAgentModel
-    )
-
-    set({
-      selectedSessionId: sessionId,
-      sessions: [newSession, ...state.sessions]
-    })
-
-    return { sessionId, messageId: userMessage.id }
-  },
-
-  // Replaces the temporary renderer id once ACP returns the real protocol session id.
-  bindPendingSession: ({ pendingSessionId, sessionId, cwd, agentFrameworkId, agentBackendId }) => {
-    if (!pendingSessionId || !sessionId) return undefined
-
-    const state = get()
-    const pendingSession = state.sessions.find(
-      (session) => session.id === pendingSessionId && session.isPending
-    )
-
-    if (!pendingSession?.activeRun) return undefined
-
-    const now = Date.now()
-
-    set({
-      selectedSessionId:
-        state.selectedSessionId === pendingSessionId ? sessionId : state.selectedSessionId,
-      sessions: state.sessions.map((session) =>
-        session.id === pendingSessionId
-          ? {
-              ...session,
-              id: sessionId,
-              isPending: false,
-              cwd: cwd ?? session.cwd,
-              agentFrameworkId: agentFrameworkId ?? session.agentFrameworkId,
-              agentBackendId: agentBackendId ?? session.agentBackendId,
-              updatedAt: now
-            }
-          : session
-      )
-    })
-
-    return {
-      sessionId,
-      messageId: pendingSession.activeRun.promptMessageId
-    }
-  },
-
-  clearPendingContextReplay: (sessionId, messageId) => {
-    set((state) => ({
-      sessions: state.sessions.map((session) =>
-        session.id === sessionId && session.pendingContextReplayMessageId === messageId
-          ? { ...session, pendingContextReplayMessageId: undefined }
-          : session
-      )
-    }))
-  },
-
+  ...createSessionMessageGraphOwner<SessionStore>(set, get),
   ...createSessionPersistenceOwner<SessionStore>(set),
 
   // Appends or extends a streamed agent message using a stable stream id.
@@ -2118,156 +1454,6 @@ export const useSessionStore = create<SessionStore>((set, get) => ({
             }
           : session
       )
-    }))
-  },
-
-  // Drops a stale interrupted turn before re-send. Removing a message advances filesRevision only if
-  // it also removes upload/artifact references from the indexed projection.
-  removeMessage: (sessionId, messageId) => {
-    if (!sessionId || !messageId) return
-
-    set((state) => ({
-      sessions: state.sessions.map((session) => {
-        if (session.id !== sessionId) return session
-        const cutIndex = session.messages.findIndex((message) => message.id === messageId)
-        if (cutIndex < 0) return session
-        const removedMessages = session.messages.slice(cutIndex)
-        const hasFiles = removedMessages.some(
-          (message) => (message.uploads?.length ?? 0) > 0 || (message.artifactIds?.length ?? 0) > 0
-        )
-        const now = Date.now()
-        const currentGraph = synchronizeSessionGraph(session, session.messages, now)
-        // Interrupted turns are immutable evidence once recorded. Fork before the unanswered prompt
-        // so retrying it cannot resurrect the old bubble, while the abandoned Branch remains available
-        // to provenance and revision history instead of being destructively rewritten.
-        const conversationGraph = forkEditedConversationMessage(
-          currentGraph,
-          messageId,
-          createConversationBranchId(),
-          now
-        )
-
-        return {
-          ...session,
-          messages: session.messages.slice(0, cutIndex),
-          conversationGraph,
-          pendingContextReplayMessageId: removedMessages.some(
-            (message) => message.id === session.pendingContextReplayMessageId
-          )
-            ? undefined
-            : session.pendingContextReplayMessageId,
-          filesRevision: hasFiles ? (session.filesRevision ?? 0) + 1 : session.filesRevision,
-          updatedAt: now
-        }
-      })
-    }))
-  },
-
-  // Drops a message and every turn after it for an edited resend. Activities are cut by timestamp
-  // because message sortIndex is transient (stripped on persist) while createdAt is durable on both
-  // sides of hydration. The kept turns are what the resend replays into the fresh agent context.
-  truncateSessionFromMessage: (sessionId, messageId) => {
-    if (!sessionId || !messageId) return
-
-    set((state) => ({
-      sessions: state.sessions.map((session) => {
-        if (session.id !== sessionId) return session
-        const cutIndex = session.messages.findIndex((message) => message.id === messageId)
-        if (cutIndex < 0) return session
-
-        const cutMessage = session.messages[cutIndex]
-        const removed = session.messages.slice(cutIndex)
-        const hasFiles = removed.some(
-          (message) => (message.uploads?.length ?? 0) > 0 || (message.artifactIds?.length ?? 0) > 0
-        )
-        const activities = session.activities?.filter(
-          (activity) => activity.createdAt < cutMessage.createdAt
-        )
-        const retainedActivityIds = new Set(activities?.map((activity) => activity.id) ?? [])
-        const activityGroups = session.activityGroups
-          ?.map((group) => ({
-            ...group,
-            activityIds: group.activityIds.filter((id) => retainedActivityIds.has(id))
-          }))
-          .filter((group) => group.activityIds.length > 0)
-        const now = Date.now()
-        const currentGraph = synchronizeSessionGraph(session, session.messages, now)
-        const conversationGraph = forkEditedConversationMessage(
-          currentGraph,
-          messageId,
-          createConversationBranchId(),
-          now
-        )
-
-        return {
-          ...session,
-          status: 'idle',
-          messages: session.messages.slice(0, cutIndex),
-          activities,
-          activityGroups,
-          conversationGraph,
-          activeRun: undefined,
-          agentStatus: undefined,
-          error: undefined,
-          errorReportable: undefined,
-          interrupted: undefined,
-          branchContextResetRequired: true,
-          pendingContextReplayMessageId: removed.some(
-            (message) => message.id === session.pendingContextReplayMessageId
-          )
-            ? undefined
-            : session.pendingContextReplayMessageId,
-          filesRevision: hasFiles ? (session.filesRevision ?? 0) + 1 : session.filesRevision,
-          updatedAt: now
-        }
-      })
-    }))
-  },
-
-  // Switches the whole durable active-path projection without contacting the Agent or Notebook.
-  activateMessageBranch: (sessionId, branchId) => {
-    if (!sessionId || !branchId) return
-    set((state) => ({
-      sessions: state.sessions.map((session) => {
-        if (
-          session.id !== sessionId ||
-          !session.conversationGraph ||
-          session.activeRun ||
-          session.status === 'running' ||
-          session.status === 'waiting-permission' ||
-          session.fixLoopActive ||
-          session.compacting ||
-          session.branchSwitchBlocked
-        ) {
-          return session
-        }
-        const activeFrame = session.conversationGraph.frames.find(
-          (frame) => frame.id === session.conversationGraph?.activeFrameId
-        )
-        if (activeFrame?.activeBranchId === branchId) return session
-        const conversationGraph = activateConversationBranch(session.conversationGraph, branchId)
-        const messages = resolveActiveConversationMessages(conversationGraph).map(
-          (message, index): ChatMessage => ({
-            ...projectConversationMessage(message),
-            sortIndex: index + 1
-          })
-        )
-        const projected = resolveActiveConversationActivities(conversationGraph)
-        return {
-          ...session,
-          conversationGraph,
-          messages,
-          activities: projected.activities.map(hydrateToolActivity),
-          activityGroups: projected.activityGroups,
-          status: 'idle',
-          activeRun: undefined,
-          error: undefined,
-          errorReportable: undefined,
-          branchContextResetRequired: true,
-          filesRevision: (session.filesRevision ?? 0) + 1,
-          updatedAt: Date.now()
-        }
-      })
     }))
   },
